updated q1

+++ Answers/Question1.cs
@@ -14,11 +14,10 @@ namespace C_Sharp_Challenge_Skeleton.Answers
             
             int result = 0;
             int n = portfolios.Length;
-            for (long i = 0; i < n-1; i++)
+            for (int i = 0; i < n-1; i++)
             {
-                for (long j = i + 1; j < n; j++)
+                for (int j = i + 1; j < n; j++)
                 {
-                    //if (portfolios[i] == portfolios[j]) continue;
 
                     int xor = portfolios[i] ^ portfolios[j];
 
@@ -28,8 +27,9 @@ namespace C_Sharp_Challenge_Skeleton.Answers
                 }
             }
             return result;
-            
 
+
+            //if (portfolios[i] == portfolios[j]) continue;
 
             /*
             int result = 0;
